Refactor service initialization in Program.cs and create Services helper class
- Moved service initialization logic from Program.cs to a new Services helper class for better organization and maintainability.
- The Services class now handles the addition of the database context and controller services, as well as the initialization of various helpers (JWT, Snowflake, PasswordHasher).
- This refactor enhances code clarity and prepares the structure for future service additions.
